TMS-Update: Type change for star count

+++ TestModule/Entity/SBEntry.cs
@@ -2,7 +2,7 @@
 {
     public class SBEntry
     {
-        public ulong StarCount { get; set; }
+        public int StarCount { get; set; }
         public ulong StarredMessageID { get; set; }
         public ulong StarredMsgChannelID { get; set; }
         public ulong SbMessageID { get; set; }
